+++ web/src/lib/components/timeline/actions/delete-asset-dialog.svelte
