+++ web/src/lib/components/timeline/base-components/base-timeline.svelte
@@ -1,16 +1,19 @@
 <script lang="ts">
-  import TimelineViewer from '$lib/components/timeline-viewer/timeline-viewer.svelte';
+  import TimelineViewer from '$lib/components/timeline/base-components/base-timeline-viewer.svelte';
   import { AssetAction } from '$lib/constants';
+  import type { DayGroup } from '$lib/managers/timeline-manager/day-group.svelte';
   import type { MonthGroup } from '$lib/managers/timeline-manager/month-group.svelte';
   import type { TimelineManager } from '$lib/managers/timeline-manager/timeline-manager.svelte';
   import type { TimelineAsset } from '$lib/managers/timeline-manager/types';
   import type { AssetInteraction } from '$lib/stores/asset-interaction.svelte';
-  import type { ScrubberListener } from '$lib/utils/timeline-util';
+  import type { ScrubberListener, TimelineYearMonth } from '$lib/utils/timeline-util';
   import type { AlbumResponseDto, PersonResponseDto } from '@immich/sdk';
   import type { Snippet } from 'svelte';
-  import Scrubber from '../shared-components/scrubber/scrubber.svelte';
+  import Scrubber from '../../shared-components/scrubber/scrubber.svelte';
 
   interface Props {
+    customThumbnailLayout?: Snippet<[TimelineAsset]>;
+
     isSelectionMode?: boolean;
     singleSelect?: boolean;
     /** `true` if this asset grid is responds to navigation events; if `true`, then look at the
@@ -31,13 +34,17 @@
     album?: AlbumResponseDto | null;
     person?: PersonResponseDto | null;
     isShowDeleteConfirmation?: boolean;
+
+    onAssetOpen?: (dayGroup: DayGroup, asset: TimelineAsset, defaultAssetOpen: () => void) => void;
     onSelect?: (asset: TimelineAsset) => void;
     onEscape?: () => void;
+
     children?: Snippet;
     empty?: Snippet;
   }
 
   let {
+    customThumbnailLayout,
     isSelectionMode = false,
     singleSelect = false,
     enableRouting,
@@ -50,6 +57,7 @@
     album = null,
     person = null,
     isShowDeleteConfirmation = $bindable(false),
+    onAssetOpen,
     onSelect = () => {},
     onEscape = () => {},
     children,
@@ -171,6 +179,7 @@
 </script>
 
 <TimelineViewer
+  {customThumbnailLayout}
   {isSelectionMode}
   {singleSelect}
   {enableRouting}
@@ -183,6 +192,7 @@
   {album}
   {person}
   {isShowDeleteConfirmation}
+  {onAssetOpen}
   {onSelect}
   {onEscape}
   {children}
