+++ web/src/lib/components/timeline/base-components/skeleton.svelte
